Add ability to send email in bulk
