+++ app/blueprints/admin/__init__.py
@@ -19,4 +19,4 @@ from flask import Blueprint
 
 bp = Blueprint("admin", __name__)
 
-from . import admin, audit, licenseseditor, tagseditor, versioneditor, warningseditor
+from . import admin, audit, licenseseditor, tagseditor, versioneditor, warningseditor, email
