+++ app/templates/admin/list.html
@@ -13,6 +13,7 @@
 				<a class="list-group-item list-group-item-action" href="{{ url_for('admin.license_list') }}">License Editor</a>
 				<a class="list-group-item list-group-item-action" href="{{ url_for('admin.version_list') }}">Version Editor</a>
 				<a class="list-group-item list-group-item-action" href="{{ url_for('admin.warning_list') }}">Warning Editor</a>
+				<a class="list-group-item list-group-item-action" href="{{ url_for('admin.send_bulk_email') }}">Send bulk email</a>
 				<a class="list-group-item list-group-item-action" href="{{ url_for('admin.switch_user') }}">Sign in as another user</a>
 			</div>
 		</div>
