+++ app/templates/admin/send_email.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %}
+
+{% block title %}
+	{{  _("Send email to %(username)s", username=user.display_name) }}
+{% endblock %}
+
+{% block content %}
+<h1>{{ self.title() }}</h1>
+
+{% from "macros/forms.html" import render_field, render_submit_field %}
+<form action="" method="POST" class="form" role="form">
+	{{ form.hidden_tag() }}
+	{{ render_field(form.subject) }}
+	{{ render_field(form.text, fieldclass="form-control markdown") }}
+	{{ render_submit_field(form.submit) }}
+</form>
+
+{% endblock %}
